+++ src/common/types.d.ts
@@ -1,10 +1,22 @@
 declare module "common-types" {
-  export type ReactRenderer = {
+  export type ReactRendererInfo = {
     id: number;
     name: string;
     version: string;
+    bundleType: RendererBundleType;
     channelId: `events:${number}`;
   };
+  export type ReactUnsupportedRendererInfo = Omit<
+    ReactRendererInfo,
+    "channelId"
+  > & {
+    reason: string;
+  };
+  export type RendererBundleType =
+    | "development"
+    | "profiling"
+    | "production"
+    | "unknown";
 
   export type FiberType = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9 | 10 | 11 | 12;
   export type FiberRootMode = 0 | 1 | 2;