+++ src/ui/components/misc/WaitingForRenderer.tsx
@@ -2,7 +2,8 @@ import * as React from "react";
 import { useReactRenderers } from "../../utils/react-renderers";
 
 export default function WaitingForRenderer() {
-  const { selected: selectedReactInstance } = useReactRenderers();
+  const { selected: selectedReactInstance, unsupportedRenderers } =
+    useReactRenderers();
 
   if (selectedReactInstance) {
     return null;
@@ -10,7 +11,23 @@ export default function WaitingForRenderer() {
 
   return (
     <div className="waiting-for-renderer">
-      Waiting for a React renderer to be connected...
+      Waiting for a supported React renderer to be connected...
+      {!unsupportedRenderers.length ? null : (
+        <div className="unsupported-renderers">
+          <div>Detected unsupported renderers:</div>
+          <ul className="unsupported-renderers__list">
+            {unsupportedRenderers.map(info => (
+              <li key={info.id}>
+                <b>
+                  {info.name} v{info.version}
+                </b>
+                {" – "}
+                {info.reason}
+              </li>
+            ))}
+          </ul>
+        </div>
+      )}
     </div>
   );
 }