+++ src/ui/utils/react-renderers.tsx
@@ -1,14 +1,17 @@
-import { ReactRenderer } from "common-types";
 import * as React from "react";
 import { remoteSubscriber } from "../rempl-subscriber";
+import { ReactRendererInfo, ReactUnsupportedRendererInfo } from "../types";
+import { RemoteProtocol } from "rempl";
 
 type ReactRenderersContext = {
-  renderers: ReactRenderer[];
-  selected: ReactRenderer | null;
+  renderers: ReactRendererInfo[];
+  unsupportedRenderers: ReactUnsupportedRendererInfo[];
+  selected: ReactRendererInfo | null;
 };
 
 const ReactRenderersContext = React.createContext<ReactRenderersContext>({
   renderers: [],
+  unsupportedRenderers: [],
   selected: null,
 });
 export const useReactRenderers = () => React.useContext(ReactRenderersContext);
@@ -17,22 +20,27 @@ export function ReactRenderersContextProvider({
 }: {
   children: React.ReactNode;
 }) {
-  const [renderers, setRenderers] = React.useState<ReactRenderer[]>([]);
+  const [{ renderers, unsupportedRenderers }, setRenderers] = React.useState<
+    RemoteProtocol["react-renderers"]["data"]
+  >({ renderers: [], unsupportedRenderers: [] });
   const value: ReactRenderersContext = React.useMemo(
     () => ({
       renderers,
+      unsupportedRenderers,
       selected: renderers[0] || null,
     }),
-    [renderers]
+    [renderers, unsupportedRenderers]
   );
 
   React.useEffect(
     () =>
-      remoteSubscriber.ns("react-renderers").subscribe(newRenderers => {
-        if (newRenderers.length) {
-          setRenderers(newRenderers);
-        }
-      }),
+      remoteSubscriber
+        .ns("react-renderers")
+        .subscribe(({ renderers, unsupportedRenderers }) => {
+          if (renderers.length || unsupportedRenderers.length) {
+            setRenderers({ renderers, unsupportedRenderers });
+          }
+        }),
     []
   );